Fix Dijkstra, add new command
Fixed Dijkstra's algorithm using locations that don't exist. Added GenerateRoutingConstraint command. Tidied up some files such as SetEdgeCosts/SetPortCosts.

+++ GoAhead/Code/Dijkstra/Node.cs
@@ -6,28 +6,38 @@ using System.Threading.Tasks;
 
 namespace GoAhead.Code.Dijkstra
 {
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <typeparam name="T"></typeparam>
     internal class Node<T>
     {
         public T Value { get; }
         public uint Index { get; }
-        public List<Tuple<Node<T>, double>> Neighbours { get; }
+        public List<Tuple<Node<T>, int>> Neighbours { get; }
 
         public Node(uint index, T value)
         {
             Value = value;
             Index = index;
-            Neighbours = new List<Tuple<Node<T>, double>>();
-            Neighbours.Add(new Tuple<Node<T>, double>(this, 0d));
+            Neighbours = new List<Tuple<Node<T>, int>>();
+            Neighbours.Add(new Tuple<Node<T>, int>(this, 0));
         }
 
-        public void AddNeighbour(Node<T> node, double distance)
+        /// <summary>
+        /// Connect a node to this node by adding the destination node and the distance as a tuple object to the list of neighbours.
+        /// </summary>
+        /// <param name="node"> The destination node to connect this node to. </param>
+        /// <param name="distance"> The distance associated with the edge. </param>
+        public void AddNeighbour(Node<T> node, int distance)
         {
-            Neighbours.Add(new Tuple<Node<T>, double>(node, distance));
+            Neighbours.Add(new Tuple<Node<T>, int>(node, distance));
         }
 
         public override string ToString()
         {
             return Value.ToString();
         }
+
     }
 }
